Move slack into storage directory
